Back_End/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymysql==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.7.0
